Fix CLI entrypoint loading
