+++ figma-diffs/package.json
@@ -18,8 +18,7 @@
   "dependencies": {
     "@figma/rest-api-spec": "0.33.0",
     "cac": "6.7.14",
-    "consola": "3.4.2",
-    "dotenv": "17.2.2"
+    "consola": "3.4.2"
   },
   "devDependencies": {
     "@types/node": "22.12.0",
