+++ texts-updater-by-figma/package.json
@@ -17,7 +17,6 @@
     "@figma/rest-api-spec": "0.33.0",
     "cac": "6.7.14",
     "consola": "3.4.2",
-    "dotenv": "16.4.7",
     "fast-levenshtein": "2.0.6",
     "prompts": "2.4.2"
   },
