+++ texts-updater-by-figma/src/cli.ts
@@ -1,7 +1,5 @@
 #!/usr/bin/env node
 
-import 'dotenv/config';
-
 import path from 'node:path';
 import { cac } from 'cac';
