--- texts-updater-by-figma/src/config/index.ts
@@ -1,3 +1,0 @@
-export * from './defaults.js';
-export * from './loader.js';
-export * from './types.js';
