--- texts-updater-by-figma/src/core/types.ts
@@ -1,2 +1,0 @@
-export type DiffPair = Record<string, string>;
-export type DiffMapping = DiffPair[];
