+++ texts-updater-by-figma/tsconfig.base.json
@@ -4,6 +4,7 @@
     "module": "esnext",
     "moduleResolution": "bundler",
     "lib": ["ES2022"],
+    "types": ["node"],
     "noEmit": true,
 
     "esModuleInterop": true,
